--- a/kolibri/content/management/commands/exportcontent.py
+++ b/kolibri/content/management/commands/exportcontent.py
@@ -1,6 +1,6 @@
 import logging as logger
 import os
-
+import sys
 from django.db.models import Sum
 from kolibri.tasks.management.commands.base import AsyncCommand
 
@@ -46,13 +46,26 @@ class Command(AsyncCommand):
                 copy = transfer.FileCopy(srcpath, dest)
 
                 with copy:
+                    if sys.platform.startswith("linux"):
+                        size = copy.fast_file_copy()
+                        if self.is_cancelled():
+                            copy.cancel()
+                        else:
+                            overall_progress_update(size)
+                            exported_files.append(dest)
 
-                    size = copy.fast_file_copy()
-                    if self.is_cancelled():
-                        copy.cancel()
                     else:
-                        overall_progress_update(size)
-                        exported_files.append(dest)
+                        with self.start_progress(total=copy.total_size) as file_cp_progress_update:
+
+                            for chunk in copy:
+                                if self.is_cancelled():
+                                    copy.cancel()
+                                    break
+                                length = len(chunk)
+                                overall_progress_update(length)
+                                file_cp_progress_update(length)
+                            else:
+                                exported_files.append(dest)
 
             if self.is_cancelled():
                 # Cancelled, clean up any already downloading files.