--- a/requirements/base.txt
+++ b/requirements/base.txt
@@ -24,3 +24,4 @@ tzlocal==1.4
 python-dateutil==2.6.0
 ifcfg==0.11b6
 sqlalchemy==1.1.12
+pyfastcopy==1.0.2