## Summary

This is to use `pyfastcopy` to fix https://github.com/learningequality/kolibri/issues/1211 with the goal that importing from USB and exporting to USB will be faster.

## Reviewer guidance

This PR will only help on speeding up the import from/export to USB on Linux systems. 


## Statistics

Update for pyfastcopy on Ubuntu 16.04:

With Python 2:
The speed for importing from USB increases **16%** with 1GB channel: 60.7129230499s --> 49.8765499592s
The speed for importing from USB increases **80%** with 6GB channel: 2926.13523316s --> 556.304387093s

The speed for exporting to USB increases **39%** with 1GB channel: 82s -->
49.5654168129s
The speed for exporting to USB increases **73%** with 6GB channel: 3367s --> 890.768079996s
    
    
With Python 3:
The speed for importing from USB increases **27%** with 1GB channel: 69.01263284683228s --> 50.56928491592407s
The speed for importing from USB with 6GB channel using `pyfastcopy` is 594.6582343578339s, while without using `pyfastcopy` is 675s at 54% (Didn't finish importing to save some time)

The speed for exporting to USB increases **39%** with 1GB channel: 77s --> 47.392226457595825s
The speed for exporting to USB with 6GB channel using `pyfastcopy` is 706.6006577014923s, while without using `pyfastcopy` is unknown (I didn't test on this one to save some time, but I assume it should be similar to the result I got for Python 2). 

> Note that the actual speed for importing from/exporting to USB should be slower than the numbers listed here, because I used a shared folder on my laptop to act as a USB driver for the VM instead of an actual USB driver, which would write files more slowly. But the relative increase should be similar. 

## Issues addressed

#1211