@lyw07 2017-10-06
It turns out that `os.sendfile`, the major implementation in `pyfastcopy`, is not available in mac os python 2 but only in python 3. So thinking about write our only file copy method to increase the speed. Closing this PR now as `pyfastcopy` is not a cross-platform tool.