@aronasorman 2017-10-11
Before I merge, I think you tested this for Mac Linux and windows right @lyw07?