@lyw07 2017-10-11
@aronasorman I only tested on Mac and Linux. So on Mac, there's no improvement for the speed of transfer. And on Linux, the improvement is listed above. I haven't tried to test on windows yet. But I can quickly set up a windows VM to test it.