@radinamatic 2017-10-11
@lyw07 Server not starting from the taskbar (Windows 10, [this installer](https://storage.googleapis.com/le-downloads/kolibri%2Frelease%2F2940%2Freport.html)), and I'm getting this error when trying from Command prompt:

```
C:\Python27\Scripts>kolibri start --debug --foreground
Traceback (most recent call last):
  File "c:\python27\lib\runpy.py", line 174, in _run_module_as_main
    "__main__", fname, loader, pkg_name)
  File "c:\python27\lib\runpy.py", line 72, in _run_code
    exec code in run_globals
  File "C:\Python27\Scripts\kolibri.exe\__main__.py", line 9, in <module>
  File "c:\python27\lib\site-packages\kolibri\utils\cli.py", line 571, in main
    initialize(debug=debug)
  File "c:\python27\lib\site-packages\kolibri\utils\cli.py", line 140, in initialize
    django.setup()
  File "c:\python27\lib\site-packages\kolibri\dist\django\__init__.py", line 18, in setup
    apps.populate(settings.INSTALLED_APPS)
  File "c:\python27\lib\site-packages\kolibri\dist\django\apps\registry.py", line 108, in populate
    app_config.import_models(all_models)
  File "c:\python27\lib\site-packages\kolibri\dist\django\apps\config.py", line 202, in import_models
    self.models_module = import_module(models_module_name)
  File "c:\python27\lib\importlib\__init__.py", line 37, in import_module
    __import__(name)
  File "c:\python27\lib\site-packages\kolibri\core\models.py", line 5, in <module>
    registry.initialize()
  File "c:\python27\lib\site-packages\kolibri\plugins\registry.py", line 80, in initialize
    plugin_module = importlib.import_module(import_string)
  File "c:\python27\lib\importlib\__init__.py", line 37, in import_module
    __import__(name)
  File "c:\python27\lib\site-packages\kolibri\plugins\coach\kolibri_plugin.py", line 4, in <module>
    from . import hooks, urls
  File "c:\python27\lib\site-packages\kolibri\plugins\coach\urls.py", line 4, in <module>
    from .api_urls import urlpatterns
  File "c:\python27\lib\site-packages\kolibri\plugins\coach\api_urls.py", line 3, in <module>
    from rest_framework import routers
  File "c:\python27\lib\site-packages\kolibri\dist\rest_framework\routers.py", line 25, in <module>
    from rest_framework import views
  File "c:\python27\lib\site-packages\kolibri\dist\rest_framework\views.py", line 18, in <module>
    from rest_framework.request import Request
  File "c:\python27\lib\site-packages\kolibri\dist\rest_framework\request.py", line 22, in <module>
    from rest_framework.settings import api_settings
  File "c:\python27\lib\site-packages\kolibri\dist\rest_framework\settings.py", line 25, in <module>
    from django.test.signals import setting_changed
  File "c:\python27\lib\site-packages\kolibri\dist\django\test\__init__.py", line 27, in <module>
    import mock  # NOQA
  File "c:\python27\lib\site-packages\kolibri\dist\mock\__init__.py", line 2, in <module>
    import mock.mock as _mock
  File "c:\python27\lib\site-packages\kolibri\dist\mock\mock.py", line 71, in <module>
    _v = VersionInfo('mock').semantic_version()
  File "c:\python27\lib\site-packages\kolibri\dist\pbr\version.py", line 461, in semantic_version
    self._semantic = self._get_version_from_pkg_resources()
  File "c:\python27\lib\site-packages\kolibri\dist\pbr\version.py", line 448, in _get_version_from_pkg_resources
    result_string = packaging.get_version(self.package)
  File "c:\python27\lib\site-packages\kolibri\dist\pbr\packaging.py", line 740, in get_version
    version = _get_version_from_git(pre_version)
  File "c:\python27\lib\site-packages\kolibri\dist\pbr\packaging.py", line 665, in _get_version_from_git
    git_dir = git._run_git_functions()
  File "c:\python27\lib\site-packages\kolibri\dist\pbr\git.py", line 131, in _run_git_functions
    if _git_is_installed():
  File "c:\python27\lib\site-packages\kolibri\dist\pbr\git.py", line 83, in _git_is_installed
    _run_shell_command(['git', '--version'])
  File "c:\python27\lib\site-packages\kolibri\dist\pbr\git.py", line 49, in _run_shell_command
    env=newenv)
  File "c:\python27\lib\subprocess.py", line 390, in __init__
    errread, errwrite)
  File "c:\python27\lib\subprocess.py", line 640, in _execute_child
    startupinfo)
TypeError: environment can only contain strings

C:\Python27\Scripts>
```