@lyw07 2017-10-11
@radinamatic thank you! Did we have some fix on Windows recently? I haven't pull from release-0.6.x branch for a long time, so I'll do that right now to get a new installer.