@radinamatic 2017-10-11
Not that I recall @lyw07, the last build I did test on Windows was #2331, and there were no startup errors, not sure if something changed in the meantime.