@lyw07 2017-10-11
Seeing the error by doing `import os; os.environ` inside subprocess.py. It shows that one of the environment variable is a unicode instead of a string. The variable is `u'KOLIBRI_HOME' : u'C:\\Users\\IEUser\\.kolibri'`