@indirectlylit 2017-10-11
> I can quickly set up a windows VM to test it

lies! ;-)