@radinamatic 2017-10-11
`Windows`, `VM` and `quickly` do not really go together...