@lyw07 2017-10-12
@aronasorman @indirectlylit @radinamatic Sorry for the late reply! I was trying to change some structure of the previous code, and that took me some time. Sorry about that and thank you for the waiting :)

So the problem I found on my previous code was that `pyfastcopy` uses a C python package called `pysendfile`, which is not cross-platform. The C code compiled on Linux in the builds couldn't be run on other platforms. Although I tested on Mac OS X before, the program I tested was originally built on Mac OS X, so I didn't realize the error until today. This `pyfastcopy` also caused the error that Radina reported above on Windows, as it was trying to import `mock`, but it didn't exist on Windows. 

What I've changed in code is to install `pyfastcopy` into `kolibri/dist/cext` folder and add its path to `sys.path` manually, as it is a C extension just like `cryptography`. However, since `pyfastcopy` is not working on Mac OS X and Windows, I've changed the code so that `pyfastcopy` will only be installed and used on Linux platforms. 

I've tested the pex file on Mac OS X, Ubuntu 16.04 and windows installer on Windows 10, doing import from network, import from/export to USB. The export to USB is working once I manually change the `block_size` to 2097152 as stated in issue #2425. Except that, everything looks good to me.

Please let me know if you find any other errors or any code implementation I can improve. Thank you!!