@lyw07 2017-10-12
This PR will only benefit linux-x86_64 platform since our builds run on this platform. The limitation of the C extension `pysendfile` makes it only accessible on the platform that compiles it. Therefore, after discussing with @aronasorman, we decide to move this to 0.7 and maintain the general approach we have currently for all the platforms on 0.6. We can handle this `pyfastcopy` the same as the other c extension `cryptography` in the future, probably having different installers for different platforms as Ben mentioned before (Not sure). But for now, closing this PR to reduce the complexity of the code that doesn't benefit the speed of transfer much may be a good idea.